Fixed failing tests
Previously, unit-tests would fail on node versions > 5. The root cause
seems to involve how `path.basename` responds when passed `undefined` as
an argument. (In older versions of node, it was treated as an empty
string. In newer versions, an error is thrown.)

I modified the affected code to pass an empty string to `path.basename`
instead of `undefined`, which resolved the problem.